+++ packages/components/src/molecules/RegionBar/RegionBar.tsx
@@ -1,7 +1,7 @@
-import type { HTMLAttributes } from 'react'
+import type { HTMLAttributes, ReactNode } from 'react'
 import React, { forwardRef } from 'react'
 
-import { Button, Icon } from '../../'
+import { Button } from '../../'
 
 export interface RegionBarProps extends HTMLAttributes<HTMLDivElement> {
   /**
@@ -12,10 +12,34 @@ export interface RegionBarProps extends HTMLAttributes<HTMLDivElement> {
    * Function called when button is clicked.
    */
   onButtonClick?: () => void
+  /**
+   * A React component that will be rendered as an icon.
+   */
+  icon?: ReactNode
+  /**
+   * Specifies a label for the location text.
+   */
+  label: string
+  /**
+   * Specifies a label for the edit text.
+   */
+  editLabel?: string
+  /**
+   * A React component that will be rendered as an icon.
+   */
+  buttonIcon?: ReactNode
 }
 
 const RegionBar = forwardRef<HTMLDivElement, RegionBarProps>(function RegionBar(
-  { postalCode, onButtonClick, ...otherProps },
+  {
+    postalCode,
+    icon,
+    label,
+    editLabel,
+    buttonIcon,
+    onButtonClick,
+    ...otherProps
+  },
   ref
 ) {
   return (
@@ -24,16 +48,16 @@ const RegionBar = forwardRef<HTMLDivElement, RegionBarProps>(function RegionBar(
         variant="tertiary"
         iconPosition="right"
         onClick={onButtonClick}
-        icon={<Icon name="CaretRight" />}
+        icon={buttonIcon}
       >
-        <Icon name="MapPin" />
+        {!!icon && icon}
         {postalCode ? (
           <>
             <span data-fs-region-bar-postal-code>{postalCode}</span>
-            <span data-fs-region-bar-cta>Edit</span>
+            {!!editLabel && <span data-fs-region-bar-cta>{editLabel}</span>}
           </>
         ) : (
-          <span data-fs-region-bar-message>Set your location</span>
+          <span data-fs-region-bar-message>{label}</span>
         )}
       </Button>
     </div>